Improve pubkey not found error handling
When the public key identified by the {Alg, KID} tuple is not found on
the IAM keystore server, it's possible to see errors like:

(node1@127.0.0.1)140> epep:jwt_decode(SampleJWT).
** exception error: no function clause matching
                    public_key:do_verify(<<"eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImtpZCI6IjIwMTcwNTIwLTAwOjAwOjAwIn0.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjEyMzIx"...>>,
                                         sha256,
                                         <<229,188,162,247,201,233,118,32,115,206,156,
                                           169,17,221,78,157,161,147,46,179,42,219,66,
                                           15,139,91,...>>,
                                         {error,not_found}) (public_key.erl, line 782)
     in function  jwtf:public_key_verify/4 (src/jwtf.erl, line 212)
     in call from jwtf:decode/3 (src/jwtf.erl, line 30)

Where the tuple {error, not_found} gets passed to the
public_key:do_verify function instead of being caught as soon as KS
returns an error.

This change throws the appropriate error message if it matches an
error returned from KS.